separated out types and enums to reduce clutter
