+++ package.json
@@ -37,6 +37,7 @@
     "husky-run": "^0.0.0"
   },
   "devDependencies": {
+    "@types/pixi.js": "^5.0.0",
     "argos-cli": "0.3.3",
     "fs-extra": "^9.1.0",
     "happo": "^5.0.0",
@@ -45,17 +46,17 @@
     "lint-staged": ">=10",
     "microbundle": "^0.13.0",
     "opener": "^1.5.1",
-    "pixi.js-legacy": "^5.3.7",
+    "pixi.js-legacy": "^5.3.8",
     "prettier": "^2.2.1",
     "standard-version": "^9.1.1",
     "typescript": "^4.1.5"
   },
   "peerDependencies": {
-    "pixi.js": "^5.3.7"
+    "pixi.js": "^5.3.8"
   },
   "scripts": {
     "demo": "npm run build && opener demo/index.html",
-    "dev": "microbundle watch --name MultiStyleText",
+    "dev": "microbundle watch",
     "build": "microbundle --name MultiStyleText",
     "prepublishOnly": "npm run build",
     "release": "standard-version",
