+++ src/InteractionEvents.ts
@@ -1,0 +1,28 @@
+export default [
+	"pointerover",
+	"pointerenter",
+	"pointerdown",
+	"pointermove",
+	"pointerup",
+	"pointercancel",
+	"pointerout",
+	"pointerleave",
+	"gotpointercapture",
+	"lostpointercapture",
+	"mouseover",
+	"mouseenter",
+	"mousedown",
+	"mousemove",
+	"mouseup",
+	"mousecancel",
+	"mouseout",
+	"mouseleave",
+	"touchover",
+	"touchenter",
+	"touchdown",
+	"touchmove",
+	"touchup",
+	"touchcancel",
+	"touchout",
+	"touchleave",
+];
