+++ src/Tags.ts
@@ -1,0 +1,9 @@
+export const TagStyle = {
+	bbcode: "bbcode",
+	xml: "xml",
+};
+
+export const TagBrackets = {
+	bbcode: ["[", "]"],
+	xml: ["<", ">"],
+};
