+++ src/types.ts
@@ -1,0 +1,148 @@
+export interface TagData {
+	name: string;
+	properties: { [key: string]: string };
+}
+
+export interface HitboxData {
+	tag: TagData;
+	hitbox: PIXI.Rectangle;
+}
+
+export type TagStyle = "bbcode" | "xml";
+export type VAlign = "top" | "middle" | "bottom" | "baseline" | number;
+
+export interface TextStyle {
+	align?: string;
+	breakWords?: boolean;
+	dropShadow?: boolean;
+	dropShadowAlpha?: number;
+	dropShadowAngle?: number;
+	dropShadowBlur?: number;
+	dropShadowColor?: string | number;
+	dropShadowDistance?: number;
+	fill?: string | string[] | number | number[] | CanvasGradient | CanvasPattern;
+	fillGradientType?: number;
+	fillGradientStops?: number[];
+	fontFamily?: string | string[];
+	fontSize?: number | string;
+	fontStyle?: string;
+	fontVariant?: string;
+	fontWeight?: string;
+	leading?: number;
+	letterSpacing?: number;
+	lineHeight?: number;
+	lineSpacing?: number;
+	lineJoin?: string;
+	miterLimit?: number;
+	padding?: number;
+	stroke?: string | number;
+	strokeThickness?: number;
+	trim?: boolean;
+	textBaseline?: string;
+	whiteSpace?: string;
+	wordWrap?: boolean;
+	wordWrapWidth?: number;
+}
+
+export interface TextStyleExtended extends TextStyle {
+	valign?: VAlign;
+	debug?: boolean;
+	tagStyle?: TagStyle;
+}
+
+export interface TextStyleExtendedWithDefault extends TextStyleExtended {
+	dropShadowAlpha?: number;
+	fillGradientStops?: number[];
+	leading?: number;
+	trim?: boolean;
+	whiteSpace?: string;
+
+	align: string;
+	breakWords: boolean;
+	dropShadow: boolean;
+	dropShadowAngle: number;
+	dropShadowBlur: number;
+	dropShadowColor: string | number;
+	dropShadowDistance: number;
+	fill: string | string[] | number | number[] | CanvasGradient | CanvasPattern;
+	fillGradientType: number;
+	fontFamily: string | string[];
+	fontSize: number | string;
+	fontStyle: string;
+	fontVariant: string;
+	fontWeight: string;
+	letterSpacing: number;
+	lineHeight: number;
+	lineSpacing: number,
+	lineJoin: string;
+	miterLimit: number;
+	padding: number;
+	stroke: string | number;
+	strokeThickness: number;
+	textBaseline: string;
+	wordWrap: boolean;
+	wordWrapWidth: number;
+
+	valign: VAlign;
+	debug: boolean;
+	tagStyle: TagStyle;
+}
+
+export type TextStyleSet = {
+	default: TextStyleExtendedWithDefault
+} & {
+	[key: string]: TextStyleExtended;
+};
+
+export interface FontProperties {
+	ascent: number;
+	descent: number;
+	fontSize: number;
+}
+
+export interface TextData {
+	text: string;
+	style: TextStyleExtended;
+	width: number;
+	height: number;
+	fontProperties: FontProperties;
+	tag: TagData;
+}
+
+export interface TextDrawingData {
+	text: string;
+	style: TextStyleExtended;
+	x: number;
+	y: number;
+	width: number;
+	ascent: number;
+	descent: number;
+	tag: TagData;
+}
+
+export interface MstDebugOptions {
+	spans: {
+		enabled?: boolean;
+		baseline?: string;
+		top?: string;
+		bottom?: string;
+		bounding?: string;
+		text?: boolean;
+	};
+	objects: {
+		enabled?: boolean;
+		bounding?: string;
+		text?: boolean;
+	};
+}
+
+export interface MstInteractionEvent extends PIXI.InteractionEvent {
+	targetTag: TagData | undefined;
+}
+export interface TextWithPrivateMembers {
+	dirty: boolean;
+	_texture: PIXI.Texture;
+	_style: PIXI.TextStyle;
+	_onTextureUpdate(): void;
+	_generateFillStyle(style: object, lines: string[]): string | number | CanvasGradient;
+}
